Rename package to com.vagujhelyigergely.calculatorm3
Update applicationId, namespace, package declarations, imports,
and proguard rules. Move source files to new directory structure.
